Convert to UTC (if needed), then drop timezone.
This should allow multi-TZ series to parse correctly and be handled by
Workbench. (I ran into an issue while testing unrelated things.) Two new
features:

* The output is never tz-aware. That helps ProcessResult handle it.
* The output is never ambiguous.

If the input text has timezone information, that information will be lost
and the user may be confused. That's better than the previous effect -- which
was that sanitizedataframe would convert back to text with mixed-TZ input.